+++ src/formal/vector_3D_s.F90
@@ -119,7 +119,7 @@ contains
      
   module procedure vector_3D_divergence
 
-    double precision, dimension(self%cells_(x_dir)+2, self%cells_(y_dir)+2, self%cells_(z_dir)+2) :: div_x_term, div_y_term, div_z_term
+    real, dimension(self%cells_(x_dir)+2, self%cells_(y_dir)+2, self%cells_(z_dir)+2) :: div_x_term, div_y_term, div_z_term
 
     call_julienne_assert(self%consistent())
